+++ src/client/actions/districtDrawing.ts
@@ -40,17 +40,22 @@ export const setSelectionTool = createAction("Set selection tool")<SelectionTool
 
 export const setGeoLevelIndex = createAction("Set geoLevel index")<number>();
 
+export const setMapLabel = createAction("Set map label")<string | undefined>();
+
 export const setGeoLevelVisibility = createAction("Set geolevel visibility")<readonly boolean[]>();
 
 export const toggleDistrictLocked = createAction("Toggle district locked")<DistrictId>();
 
 export const toggleLimitDrawingToWithinCounty = createAction("Limit drawing to within county")();
 
+export const showKeyboardShortcutsModal = createAction("Show keyboard shortcuts modal")<boolean>();
+
 export const showAdvancedEditingModal = createAction("Show advanced editing warning modal")<
   boolean
 >();
 
 export const showCopyMapModal = createAction("Show copy map modal")<boolean>();
+export const setImportFlagsModal = createAction("Show import flags modal")<boolean>();
 
 export const undo = createAction("Undo project action")();
 export const redo = createAction("Redo project action")();