+++ src/client/reducers/districtDrawing.ts
@@ -19,6 +19,7 @@ import {
   setSelectionTool,
   showAdvancedEditingModal,
   showCopyMapModal,
+  setImportFlagsModal,
   toggleDistrictLocked,
   undo,
   replaceSelectedGeounits,
@@ -31,7 +32,9 @@ import {
   FindTool,
   toggleLimitDrawingToWithinCounty,
   selectEvaluationMetric,
-  setZoomToDistrictId
+  setZoomToDistrictId,
+  setMapLabel,
+  showKeyboardShortcutsModal
 } from "../actions/districtDrawing";
 import { updateDistrictsDefinition, updateDistrictLocks } from "../actions/projectData";
 import { SelectionTool } from "../actions/districtDrawing";
@@ -110,6 +113,8 @@ export interface DistrictDrawingState {
   readonly selectionTool: SelectionTool;
   readonly showAdvancedEditingModal: boolean;
   readonly showCopyMapModal: boolean;
+  readonly showKeyboardShortcutsModal: boolean;
+  readonly showImportFlagsModal: boolean;
   readonly findMenuOpen: boolean;
   readonly evaluateMode: boolean;
   readonly evaluateMetric: EvaluateMetric | undefined;
@@ -118,6 +123,7 @@ export interface DistrictDrawingState {
   readonly limitSelectionToCounty: boolean;
   readonly saving: SavingState;
   readonly undoHistory: UndoHistory;
+  readonly mapLabel: string | undefined;
 }
 
 export const initialDistrictDrawingState: DistrictDrawingState = {
@@ -128,12 +134,15 @@ export const initialDistrictDrawingState: DistrictDrawingState = {
   selectionTool: SelectionTool.Default,
   showAdvancedEditingModal: false,
   showCopyMapModal: false,
+  showImportFlagsModal: false,
+  showKeyboardShortcutsModal: false,
   findMenuOpen: false,
   evaluateMode: false,
   evaluateMetric: undefined,
   findTool: FindTool.Unassigned,
   limitSelectionToCounty: false,
   saving: "unsaved",
+  mapLabel: "undefined",
   undoHistory: {
     past: [],
     present: {
@@ -174,6 +183,11 @@ const districtDrawingReducer: LoopReducer<ProjectState, Action> = (
       return {
         ...state,
         zoomToDistrictId: action.payload
+      };
+    case getType(setMapLabel):
+      return {
+        ...state,
+        mapLabel: action.payload
       };
     case getType(addSelectedGeounits):
       return loop(
@@ -271,10 +285,20 @@ const districtDrawingReducer: LoopReducer<ProjectState, Action> = (
         ...state,
         showAdvancedEditingModal: action.payload
       };
+    case getType(showKeyboardShortcutsModal):
+      return {
+        ...state,
+        showKeyboardShortcutsModal: action.payload
+      };
     case getType(showCopyMapModal):
       return {
         ...state,
         showCopyMapModal: action.payload
+      };
+    case getType(setImportFlagsModal):
+      return {
+        ...state,
+        showImportFlagsModal: action.payload
       };
     case getType(toggleFind):
       return {