+++ src/server/src/project-templates/services/project-templates.service.ts
@@ -20,7 +20,6 @@ export class ProjectTemplatesService extends TypeOrmCrudService<ProjectTemplate>
       .innerJoinAndSelect("projectTemplate.projects", "projects")
       .innerJoinAndSelect("projects.user", "user")
       .where("organization.slug = :slug", { slug: slug })
-      .andWhere("projectTemplate.isActive = TRUE")
       .select([
         "projectTemplate.name",
         "projectTemplate.numberOfDistricts",
@@ -48,7 +47,6 @@ export class ProjectTemplatesService extends TypeOrmCrudService<ProjectTemplate>
       .leftJoinAndSelect("projectTemplate.projects", "projects", "projects.isFeatured = TRUE")
       .innerJoinAndSelect("projects.user", "user")
       .where("organization.slug = :slug", { slug: slug })
-      .andWhere("projectTemplate.isActive = TRUE")
       .select([
         "projectTemplate.name",
         "projectTemplate.numberOfDistricts",